[#121345079] Add role to separate user and admin

+++ app/controllers/photos_controller.rb
@@ -4,6 +4,7 @@ class PhotosController < ApplicationController
   def destroy
     restaurant = Restaurant.find(params[:restaurant_id])
     photo = restaurant.photos.find(params[:id])
+    authorize photo
     photo.destroy
     redirect_to restaurant_path(restaurant), info: "Success"
   end
